Progress on linking
Linking should remove the need for the CPool everywhere.  Recursively find all XXXCon items and add a cached XXXPart instead.  This search crosses CPool borders in some cases and thus requires the full Repo.  Handles interface method lookups, but not globals like toString.
All the XXXCon classes cache something after linking, so a 2nd call to link does nothing (as a form of visit bit).  The XXXPart classes use an external visit bit to avoid double-linking.
Hide more of the XXXCon fields and use the XXXPart fields instead.
Replace more StringCon with interned String.
MethodPart is now a MMEthodPart with more stuff; MMethods retain the notion of a collection of Methods, and both can be stored in a MMethod array.