Updated Windows launcher to respect XDK_HOME env var (#177)
As Marcus points out, there's still future work to be done here on documentation and build automation (and reproducibility on other build machines). Also, there's a block of good code in the Windows specific .c file that I had to comment out and replace with the old (considered potentially "unsafe", but still widely used) POSIX API call for accessing env vars -- a function that I was trying to avoid. For some reason, the newer POSIX API call wasn't found, even though it's documented by (drum roll) Microsoft, so my assumption is that there's a disconnect somewhere between the MinGW packages (supporting Windows with gcc) and the Windows libraries.