Optimized docker file and ci logic, fixed some issues with too long runs. (#276)
Optimized docker file and ci logic, fixed some issues with too long runs. Slimmed down dockerfile a bit, but not to distroless yet, since that can execute binary launchers.  (#276)

We should probably just run the jar directly in a distroless java for minimum size. Keeping those binary launchers around just adds complexity everywhere and it is unclear if they can't just be a script, since the tty issues on windows that were the original motivation seem to be solvable anyway.